refactor: replace mock data with real metrics across the dashboard
Removes lib/mock-data.ts and the pipelines demo page/components, and wires
every chart, hook, and route to honest empty states instead of fallback
mock data. Also fixes timeseries bucket zero-filling and per-minute rate
normalization, widens the traces window to match the time-range picker,
and refreshes project access checks to query membership live instead of
trusting a cached session list.

+++ apps/api/api/schemas.py
@@ -52,10 +52,10 @@ class ProjectResponse(BaseModel):
 
 class TimeseriesPoint(BaseModel):
     time: str
-    requests: int
-    errors2xx: int
-    errors4xx: int
-    errors5xx: int
+    requests: float   # per-minute rate, normalized by bucket width — never a raw bucket count
+    errors2xx: float
+    errors4xx: float
+    errors5xx: float
     p99: float
